Recommit clean code without secrets

+++ be/src/main/java/com/q2k/meditech/dto/AssignRolesDTO.java
@@ -20,5 +20,5 @@ import java.util.Set;
 public class AssignRolesDTO {
     
     @NotEmpty(message = "Role IDs cannot be empty")
-    private Set<Integer> roleIds;
+    private Set<Long> roleIds;
 }
